// src/lib/types/accounts.ts
export interface Account{
    id: string;
    title: string;
    firstName: string;
    lastName: string;
    email: string;
    role: string;
    createdAt: string;
    updatedAt: string;
}

export interface TableAccountProps {
    id: string;
    title: string;
    firstName: string;
    lastName: string;
    email: string;
    role: string;
    createdAt: string;
}